api/pyproject.toml
[project]
name = "ams-dashboard-api"
version = "0.1.0"
description = "AMS Dashboard API (Django)"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15",
    "django-filter>=24.3",
    "django-environ>=0.11",
    "psycopg[binary]>=3.2",
    "httpx>=0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-django>=4.9",
    "ruff>=0.7",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ams_dashboard.settings"
python_files = ["test_*.py", "tests.py"]
